tools/gdrive-dataset-manager/pyproject.toml
[project]
name = "gdrive-dataset-manager"
version = "1.0.0"
description = "Professional CLI tool for managing AI datasets on Google Drive"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Pixelated Empathy", email = "pixelatedempathy@github.com"}
]
maintainers = [
    {name = "Pixelated Empathy", email = "pixelatedempathy@github.com"}
]
keywords = ["google-drive", "dataset", "cli", "ai", "machine-learning"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Archiving",
    "Topic :: Utilities",
]
requires-python = ">=3.11"
dependencies = [
    "click>=8.2.1",
    "colorama>=0.4.6",
    "google-api-python-client>=2.181.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
    "tqdm>=4.67.1",
]

[project.scripts]
gdrive-dataset-manager = "gdrive_dataset_manager.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gdrive_dataset_manager"]
